changed position of emai_text_form_field, password_text_form_field, moved out \n condition form_header
